Dictionary App JAVASCRIPT PROJECT
This project is a simple dictionary application that uses JavaScript to fetch word meanings and pronunciations from an API. Users can type a word into the search bar to learn its meaning and pronunciation. The application uses an API to fetch word definitions and pronunciations, allowing users to search for words.